Change LastMember to cooperate better with rigid vars and local dicts
The use of a closed type family works well when the entire effect list
is known, but when it’s quantified over, the closed type family gets
stuck (since the effect list is rigid). Sometimes this is okay, since it
simply means the LastMember constraint is insoluble, but can be
satisfied if an explicit constraint brings a local LastMember dictionary
into scope. Performing the recursion directly in the instance avoids
this problem, though it does require overlapping instances.

fixes #6